Remove problematic dynamic imageicon + optimize scale function
I was getting a scale feedback loop where the dynamic image was increasing infinitely since the panel resized since the image resized since the panel resized, etc... Removed since it was beyond the scope of this project, and I need to get shit done. Static images are ok for now.